DNM: Just a PoC to prove it works
Depends on https://github.com/mandre/baremetal-runtimecfg/tree/render-vipless

The right solution would be to create a new command in
baremetal-runtimecfg for checking if the node is VIPable and call this
new command from a new service running on the node before kubelet
starts, similar to what we do with nodeip-configuration.service.

The new service must delete the keepalived manifest when the node is not
VIPable.